+++ src/canvas-effects/effects-layer.js
@@ -58,9 +58,9 @@ export class SequencerInterfaceLayer extends InteractionLayer {
         new PIXI.Graphics()
       );
 
-      this.suggestionPoint.filters = [new PIXI.filters.AlphaFilter(0.75)];
-      this.effectSourcePosition.filters = [new PIXI.filters.AlphaFilter(0.75)];
-      this.effectTargetPosition.filters = [new PIXI.filters.AlphaFilter(0.75)];
+      this.suggestionPoint.filters = [new PIXI.AlphaFilter(0.75)];
+      this.effectSourcePosition.filters = [new PIXI.AlphaFilter(0.75)];
+      this.effectTargetPosition.filters = [new PIXI.AlphaFilter(0.75)];
 
       this.effectSelectionBorder.zIndex = 1;
 
@@ -467,12 +467,15 @@ export class SequencerAboveUILayer {
   }
 
   static getLayer() {
-    return layer ? layer.app.stage : canvas.uiEffectsLayer;
+    return layer ? layer.app.stage : canvas.sequencerEffectsUILayer;
   }
 
   static addChild(...args) {
-    const result = this.getLayer().addChild(...args);
-    layer.app.stage.renderable = layer.app.stage.children.length > 0;
+		const layer = this.getLayer();
+    const result = layer.addChild(...args);
+		if (layer.app?.stage) {
+			layer.renderable = layer.children.length > 0;
+		}
     return result;
   }
 
@@ -481,10 +484,11 @@ export class SequencerAboveUILayer {
   }
 
   static removeContainerByEffect(inEffect) {
-    const child = this.getLayer().children.find((child) => child === inEffect);
+		const layer = this.getLayer();
+    const child = layer.children.find((child) => child === inEffect);
     if (!child) return;
-    this.getLayer().removeChild(child);
-    layer.app.stage.renderable = layer.app.stage.children.length > 0;
+    layer.removeChild(child);
+    layer.renderable = layer.children.length > 0;
   }
 
   updateTransform() {